fix mac-specific keybinding sync
The keybindings file was always being set to keybindings.json because
of missing braces. In addition, there were some inconsistencies
between files on what the name of the Mac keybinding filename was.
keybindingDefault is removed from the creation snippet as it should
just be using keybindings.json. Lastly, to prevent settings overrides,
we no longer clear files from the gist when the filename begins with
"keybindings".