Added support for "Code - OSS"
Looks like `environmentPath.ts` is looking for a specific path "/Code" ( or "/Code - Insiders", no idea what is it though). For me, this path doesn't exist since I'm using the open-source version of VS Code, thus the path which should be looked for is "/Code - OSS". I don't know if any other changes in other files should be made, but this seems to work.