#754 and #865
* Fix #754

The pragma util keeps a copy of the settings that are upload to the
gist. The file is located at <context.globalStoragePath>/settings.sync.

If the content to be upload doesn't deffer from the local content after
taking out the @sync-ignore, the file wont be uploaded.

* Fix #754: Sync object

Checks the global storage path at constructor.
Checks if the content of the setting file should be uploaded.

* Update tests

* Fix #865

* Fix broken tests

* Fixing Octokit wrong type for GistsResponse

'file' field is wrong with static data.

* Defining PragmaUtil behavior

Pragma util should only parse string content, checking each line for pragma statements.

* Updating test for Pragma definition

* Sync: don't upload is not necessary

As the extension gets the entire gists before commiting an upload, we can check each file of the gists against our local files in order to determine if the content has changed.

There is no need to keel a local copy.