Make IsGistNewer compare against latestDownload
Instead of comparing the gist's latestUpload time against the local
latestUpload time, this function should compare the gist time against
the local latestDownload time in order to determine if the cloud
settings are newer than the local settings.

Also, on an upload, set the local latestDownload time to the same
time as the upload since the settings are effectively "downloaded" on
an upload.

This protects against the following scenario:

- Machine A uploads settings to cloud
- Machine B downloads these settings
- Machine B changes settings and uploads them to the cloud
- Machine A downloads the settings
- Machine A changes settings and tries to upload them

The last upload attempt fails because A's last upload time is before
the cloud's last upload time, even though A downloaded the settings.
A forced upload is needed in order to upload the settings.
It therefore makes sense to compare against the local download time.