Fix changed files check to work with snippets
When checking for the existance of files in the gist, use the
settings file "gistName" to lookup the file in the gist instead of
"fileName" in order to find any files in a subdir.  The gist uses "|"
as a separator in the "gistName" for files in a subdir and so using
a plain filename does not match.

This fixes the case where a user has snippets and does an upload with
no changes to settings or snippets.  The snippets are already up in the
gist.  The code that checks for files that are local and not in the gist
fails to match the fileName to the gistName and incorrectly concludes
that the file is not in the gist.  This causes the upload code to upload
all the files even if there were no changes, resulting in an "empty"
gist revision with only the upload timestamp changed.