Always use "|" as a path sep for GistNames
Without this fix, the distObject that is prepared for the PATCH
command sent to GitHub can contain malformed objects resulting in
a HTTP 422 error.

For example, if you have a c.json snippet, the distObject will
contain a "snippets/c.json" file with valid contents and also a
"snippets|c.json" file with null content.