Add forceUpdate dialog for no lastDownload case
After resetting the settings, there is no lastDownload date
stored locally, so we can't tell how the gist settings relate
to the local settings.  So, present the user with a general
"Do you want to force upload?" dialog.

This new dialog differs from the "gistNewer" dialog because we
can definately say that the gist settings changed since the last
download in the "gistNewer" case, which is useful to the user.