Upload UX Improvements (#1042)
* Always use "|" as a path sep for GistNames

Without this fix, the distObject that is prepared for the PATCH
command sent to GitHub can contain malformed objects resulting in
a HTTP 422 error.

For example, if you have a c.json snippet, the distObject will
contain a "snippets/c.json" file with valid contents and also a
"snippets|c.json" file with null content.

* Add status bar msg for upload got latest

Add a status bar message for the case when the local files already
match those in the gist when the user requests an upload.

This is a message that is symmetrical to the message shown when
a user tries to download and already has the latest.

* Streamline Upload User Experience

- Remove "Don't show this again" logic and setting from the
  "Gist is newer than what you downloaded dialog".  This is not
  needed after the following change.
- Change the "Don't show this again" dialog to ask if the user
  wants to do a one-time force upload if the gist last upload time
  is prior to the local last download time.
- Fix bug where the user had to resort to a force upload when
  trying to upload settings changes.
- NLS for new messages and dialogs.
- README.md addition to explain how upload conflicts are handled.

* Apply suggestions from code review

Fix Chinese nls packages

Co-Authored-By: YunChaoTsai <yunchao2081@gmail.com>

* Preserve possible null in call to IsGistNewer

Move the conversion of customSettings.lastDownload to a Date
to inside of IsGistNewer().  This allows the function to check
for a null value for customSettings.lastDownload before converting.
Converting a null results in a valid Date of Jan 1, 1970 which
resulted in always comparing against this old date.

* Don't show "Gist is newer" dialog if force upload

When forceUpload is true, we should skip the dialog warning user
that the gist is newer than the last download.

* Remove trailing whitespace

* Fix NLS-related button value check

* Add forceUpdate dialog for no lastDownload case

After resetting the settings, there is no lastDownload date
stored locally, so we can't tell how the gist settings relate
to the local settings.  So, present the user with a general
"Do you want to force upload?" dialog.

This new dialog differs from the "gistNewer" dialog because we
can definately say that the gist settings changed since the last
download in the "gistNewer" case, which is useful to the user.